backend/requirements.txt
google-generativeai==0.8.3
pinecone-client==5.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
pypdf==5.1.0
python-docx==1.1.2
tiktoken==0.8.0
python-dotenv==1.0.1
email-validator==2.2.0
huggingface_hub==0.27.0
